# feat(router): Add pprof handlers for performance profiling