--- a/.env.development
+++ b/.env.development
@@ -3,6 +3,8 @@ CLIENT_URL=http://localhost:3000
 
 SERVER_ADDRESS=localhost:8080
 
+PROFILER_ADDRESS=localhost:2080
+
 FILE_STORAGE_PATH=store.json
 DATABASE_DSN=postgres://postgres:postgres@database:5432/shortly-development?sslmode=disable