--- a/.env.test
+++ b/.env.test
@@ -3,6 +3,8 @@ CLIENT_URL=http://localhost:3000
 
 SERVER_ADDRESS=localhost:8080
 
+PROFILER_ADDRESS=localhost:2080
+
 FILE_STORAGE_PATH=store-test.json
 DATABASE_DSN=postgres://postgres:postgres@localhost:5432/shortly-test?sslmode=disable