--- a/.github/workflows/integrations.yaml
+++ b/.github/workflows/integrations.yaml
@@ -177,20 +177,8 @@ jobs:
               -database-dsn='postgres://postgres:postgres@localhost:5432/shortly-test?sslmode=disable'
 
       - name: "Code increment #16"
-        if: |
-          github.ref == 'refs/heads/main' ||
-          github.head_ref == 'iter16' ||
-          github.head_ref == 'iter17' ||
-          github.head_ref == 'iter18' ||
-          github.head_ref == 'iter19' ||
-          github.head_ref == 'iter20' ||
-          github.head_ref == 'iter21' ||
-          github.head_ref == 'iter22' ||
-          github.head_ref == 'iter23' ||
-          github.head_ref == 'iter24'
         run: |
-          shortenertestbeta -test.v -test.run=^TestIteration16$ \
-              -source-path=. \
+          shortenertestbeta -test.v -test.run=^TestIteration16$ -source-path=.
 
       - name: "Code increment #17"
         if: |