--- a/internal/app/config/config_test.go
+++ b/internal/app/config/config_test.go
@@ -40,6 +40,7 @@ func Test_LoadConfig(t *testing.T) {
 				Addr:            "localhost:8080",
 				BaseURL:         "http://localhost:8080",
 				ClientURL:       "http://localhost:3000",
+				ProfilerAddr:    "localhost:2080",
 				FileStoragePath: "store-test.json",
 				DatabaseDSN:     "postgres://postgres:postgres@localhost:5432/shortly-test?sslmode=disable",
 				SecretKey:       "jwt-secret-key",
@@ -51,6 +52,7 @@ func Test_LoadConfig(t *testing.T) {
 				"-a", "localhost:5000",
 				"-b", "http://localhost:5000",
 				"-c", "http://localhost:6000",
+				"-p", "localhost:2080",
 				"-f", "store-test.json",
 				"-d", "postgres://postgres:postgres@localhost:5432/shortly-test?sslmode=disable",
 				"-s", "jwt-secret-key",
@@ -59,6 +61,7 @@ func Test_LoadConfig(t *testing.T) {
 				"SERVER_ADDRESS":    "localhost:3000",
 				"BASE_URL":          "http://localhost:3000",
 				"CLIENT_URL":        "http://localhost:6000",
+				"PROFILER_ADDRESS":  "localhost:2080",
 				"FILE_STORAGE_PATH": "store-test.json",
 				"DATABASE_DSN":      "postgres://postgres:postgres@localhost:5432/shortly-test?sslmode=disable",
 				"SECRET_KEY":        "jwt-secret-key",
@@ -68,6 +71,7 @@ func Test_LoadConfig(t *testing.T) {
 				Addr:            "localhost:3000",
 				BaseURL:         "http://localhost:3000",
 				ClientURL:       "http://localhost:6000",
+				ProfilerAddr:    "localhost:2080",
 				FileStoragePath: "store-test.json",
 				DatabaseDSN:     "postgres://postgres:postgres@localhost:5432/shortly-test?sslmode=disable",
 				SecretKey:       "jwt-secret-key",
@@ -88,6 +92,7 @@ func Test_LoadConfig(t *testing.T) {
 			assert.Equal(t, tt.expected.Addr, result.Addr)
 			assert.Equal(t, tt.expected.BaseURL, result.BaseURL)
 			assert.Equal(t, tt.expected.ClientURL, result.ClientURL)
+			assert.Equal(t, tt.expected.ProfilerAddr, result.ProfilerAddr)
 			assert.Equal(t, tt.expected.FileStoragePath, result.FileStoragePath)
 			assert.Equal(t, tt.expected.DatabaseDSN, result.DatabaseDSN)
 			assert.Equal(t, tt.expected.SecretKey, result.SecretKey)