--- a/internal/app/server/pprof.go
+++ b/internal/app/server/pprof.go
@@ -1,6 +1,7 @@
 package server
 
 import (
+	"context"
 	"net/http"
 	"net/http/pprof"
 	"time"
@@ -10,7 +11,16 @@ import (
 	"shortly/internal/app/config"
 )
 
-func NewPprofServer(cfg *config.Config) *http.Server {
+type PprofServer interface {
+	Run() error
+	Shutdown(ctx context.Context) error
+}
+
+type pprofServer struct {
+	httpServer *http.Server
+}
+
+func NewPprofServer(cfg *config.Config) PprofServer {
 	handler := chi.NewRouter()
 
 	handler.HandleFunc("/debug/pprof/", pprof.Index)
@@ -20,9 +30,19 @@ func NewPprofServer(cfg *config.Config) *http.Server {
 	handler.HandleFunc("/debug/pprof/trace", pprof.Trace)
 	handler.HandleFunc("/debug/pprof/heap", pprof.Handler("heap").ServeHTTP)
 
-	return &http.Server{
-		Addr:        cfg.ProfilerAddr,
-		Handler:     handler,
-		ReadTimeout: 60 * time.Second,
+	return &pprofServer{
+		httpServer: &http.Server{
+			Addr:        cfg.ProfilerAddr,
+			Handler:     handler,
+			ReadTimeout: 60 * time.Second,
+		},
 	}
+}
+
+func (p *pprofServer) Run() error {
+	return p.httpServer.ListenAndServe()
+}
+
+func (p *pprofServer) Shutdown(ctx context.Context) error {
+	return p.httpServer.Shutdown(ctx)
 }