Добавьте в свой проект бенчмарки, измеряющие скорость выполнения важнейших компонентов вашей системы.
Проведите анализ использования памяти вашим проектом, определите и исправьте неэффективные части кода по следующему алгоритму:

- Используя профилировщик `pprof`, сохраните профиль потребления памяти вашим проектом в директорию `profiles` с именем `base.pprof`
- Изучите полученный профиль, определите и исправьте неэффективные части вашего кода
- Повторите пункт 1 и сохраните новый профиль потребления памяти в директорию `profiles` с именем `result.pprof`

Проверьте результат внесённых изменений командой:

```
pprof -top -diff_base=profiles/base.pprof profiles/result.pprof
```


```
wrk -t16 -c100 -d30s -s scripts/benchmark.lua http://localhost:8080/api/shorten
Running 30s test @ http://localhost:8080/api/shorten
  16 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     9.34ms    7.74ms 265.83ms   97.62%
    Req/Sec   674.53     76.45     0.93k    73.34%
  322242 requests in 30.05s, 115.86MB read
Requests/sec:  10722.89
Transfer/sec:      3.86MB
```